Add `/dev/note` endpoint for live debugging
The server can take time to reflect code changes, and if changes do not
affect the UI, it can be difficult to tell whether the server has
updated. This is especially troublesome when trying many potential fixes
for a bug without knowing that any will work.

To fix this problem, this commit adds a /dev/note endpoint that serves
up a custom "note" developers can change whenever they want to check
that the server has updated. In the future, perhaps debugging features
like this should be authenticated in some way.