Refactor execute_insert to wrap execute_mod
The `execute_insert` function in `apiutils` actually generalizes to
executing any statement that modifies the database without returning
data from it. To reflect this in the naming, rename `execute_insert` to
`execute_mod` and then have a function called `execute_insert` wrap
`execute_mod` to avoid the headache of changing all usages of
`execute_insert`.